*: cleanup util and engine components
