+++ components/engine/Cargo.toml
@@ -20,6 +20,11 @@ slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global
 time = "0.1"
 sys-info = "0.5.1"
 tikv_alloc = { path = "../tikv_alloc", default-features = false }
+serde = "1.0"
+serde_derive = "1.0"
+toml = "0.4"
+hex = "0.3"
+tikv_util = { path = "../tikv_util" }
 
 [dependencies.prometheus]
 version = "0.4.2"
