+++ components/tikv_util/Cargo.toml
@@ -39,9 +39,7 @@ serde_json = "1.0"
 serde_derive = "1.0"
 grpcio = { version = "0.4", features = [ "secure" ] }
 crossbeam = "0.5"
-hex = "0.3"
 panic_hook = { path = "../panic_hook" }
-engine = { path = "../engine" }
 
 [dependencies.prometheus]
 version = "0.4.2"
