+++ src/bin/tikv-server.rs
@@ -39,6 +39,7 @@ use fs2::FileExt;
 
 use engine::rocks;
 use engine::rocks::util::metrics_flusher::{MetricsFlusher, DEFAULT_FLUSHER_INTERVAL};
+use engine::rocks::util::security::encrypted_env_from_cipher_file;
 use engine::Engines;
 use tikv::config::{check_and_persist_critical_config, TiKvConfig};
 use tikv::coprocessor;
@@ -52,7 +53,7 @@ use tikv::server::status_server::StatusServer;
 use tikv::server::transport::ServerRaftStoreRouter;
 use tikv::server::{create_raft_storage, Node, Server, DEFAULT_CLUSTER_ID};
 use tikv::storage::{self, AutoGCConfig, DEFAULT_ROCKSDB_SUB_DIR};
-use tikv_util::security::{self, SecurityManager};
+use tikv_util::security::SecurityManager;
 use tikv_util::time::Monitor;
 use tikv_util::worker::{Builder, FutureWorker};
 use tikv_util::{self as tikv_util, check_environment_variables};
@@ -146,7 +147,7 @@ fn run_raft_server(pd_client: RpcClient, cfg: &TiKvConfig, security_mgr: Arc<Sec
 
     // Create encrypted env from cipher file
     let encrypted_env = if !cfg.security.cipher_file.is_empty() {
-        match security::encrypted_env_from_cipher_file(&cfg.security.cipher_file, None) {
+        match encrypted_env_from_cipher_file(&cfg.security.cipher_file, None) {
             Err(e) => fatal!(
                 "failed to create encrypted env from cipher file, err {:?}",
                 e
