+++ src/config.rs
@@ -28,14 +28,13 @@ use crate::storage::mvcc::properties::{
     MvccPropertiesCollectorFactory, RangePropertiesCollectorFactory,
 };
 use crate::storage::{Config as StorageConfig, DEFAULT_ROCKSDB_SUB_DIR};
+use engine::rocks::util::config::{self as rocks_config, CompressionType};
 use engine::rocks::util::{
     db_exist, CFOptions, EventListener, FixedPrefixSliceTransform, FixedSuffixSliceTransform,
     NoopSliceTransform,
 };
 use engine::{CF_DEFAULT, CF_LOCK, CF_RAFT, CF_WRITE};
-use tikv_util::config::{
-    self, compression_type_level_serde, CompressionType, ReadableDuration, ReadableSize, GB, KB, MB,
-};
+use tikv_util::config::{self, ReadableDuration, ReadableSize, GB, KB, MB};
 use tikv_util::security::SecurityConfig;
 use tikv_util::time::duration_to_sec;
 
@@ -124,7 +123,7 @@ macro_rules! cf_config {
             pub bloom_filter_bits_per_key: i32,
             pub block_based_bloom_filter: bool,
             pub read_amp_bytes_per_bit: u32,
-            #[serde(with = "compression_type_level_serde")]
+            #[serde(with = "rocks_config::compression_type_level_serde")]
             pub compression_per_level: [DBCompressionType; 7],
             pub write_buffer_size: ReadableSize,
             pub max_write_buffer_number: i32,
@@ -135,12 +134,12 @@ macro_rules! cf_config {
             pub level0_slowdown_writes_trigger: i32,
             pub level0_stop_writes_trigger: i32,
             pub max_compaction_bytes: ReadableSize,
-            #[serde(with = "config::compaction_pri_serde")]
+            #[serde(with = "rocks_config::compaction_pri_serde")]
             pub compaction_pri: CompactionPriority,
             pub dynamic_level_bytes: bool,
             pub num_levels: i32,
             pub max_bytes_for_level_multiplier: i32,
-            #[serde(with = "config::compaction_style_serde")]
+            #[serde(with = "rocks_config::compaction_style_serde")]
             pub compaction_style: DBCompactionStyle,
             pub disable_auto_compactions: bool,
             pub soft_pending_compaction_bytes_limit: ReadableSize,
@@ -572,7 +571,7 @@ impl TitanDBConfig {
 #[serde(default)]
 #[serde(rename_all = "kebab-case")]
 pub struct DbConfig {
-    #[serde(with = "config::recovery_mode_serde")]
+    #[serde(with = "rocks_config::recovery_mode_serde")]
     pub wal_recovery_mode: DBRecoveryMode,
     pub wal_dir: String,
     pub wal_ttl_seconds: u64,
@@ -590,7 +589,7 @@ pub struct DbConfig {
     pub info_log_keep_log_file_num: u64,
     pub info_log_dir: String,
     pub rate_bytes_per_sec: ReadableSize,
-    #[serde(with = "config::rate_limiter_mode_serde")]
+    #[serde(with = "rocks_config::rate_limiter_mode_serde")]
     pub rate_limiter_mode: DBRateLimiterMode,
     pub auto_tuned: bool,
     pub bytes_per_sync: ReadableSize,
@@ -794,7 +793,7 @@ impl RaftDefaultCfConfig {
 #[serde(default)]
 #[serde(rename_all = "kebab-case")]
 pub struct RaftDbConfig {
-    #[serde(with = "config::recovery_mode_serde")]
+    #[serde(with = "rocks_config::recovery_mode_serde")]
     pub wal_recovery_mode: DBRecoveryMode,
     pub wal_dir: String,
     pub wal_ttl_seconds: u64,
